Class Today
I might not be in class today.  I had a utility problem (i didn't pay it ) and i have to wait for them to cut my water back on.  When the guy gets here i will leave and make it in as soon as i can.  Sorry for the Inconvenience but I will still be here working and fixing my push problems.